Add files via upload
Here's a brief description of each of these functions:

AVERAGE: This function calculates the average (arithmetic mean) of all the numbers in a column. It is straightforward and works with columns containing numeric values. It ignores any non-numeric values in the column.

Example: AVERAGE([Column])

AVERAGEA: Unlike AVERAGE, this function calculates the average (arithmetic mean) of the values in a column but handles text and non-numeric values as well. It treats text and other non-numeric values as 0 while calculating the average.

Example: AVERAGEA([Column])

AVERAGEX: This function is more versatile. It calculates the average (arithmetic mean) of a set of expressions evaluated over a table. This means you can use it to perform complex calculations on a table and then find the average of those results. It's especially useful for scenarios where you want to calculate averages based on more advanced criteria.

Example: AVERAGEX(Table, [Expression])